automation/worker/requirements.txt
fastapi
uvicorn
requests
prometheus_fastapi_instrumentator
nats-py
aiohttp
munch